Atividade_03/ex_sprint_5/pyproject.toml
[project]
name = "ex-sprint-5"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.123.10",
    "httpx>=0.28.1",
    "passlib>=1.7.4",
    "pyjwt>=2.10.1",
    "pytest>=9.0.1",
    "sqlalchemy>=2.0.44",
]
